pyproject.toml
[tool.poetry]
name = "tap-amazon-seller"
version = "0.0.2"
description = "`tap-amazon-seller` is a Singer tap for Amazon-Seller, built with the Meltano SDK for Singer Taps."
authors = ["Adil Ahmed"]
keywords = [
    "ELT",
    "Amazon-Seller",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
requests = "^2.25.1"
singer-sdk = "^0.4.4"
python-amazon-sp-api = "^0.13.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_amazon_seller"

[build-system]
# Uncomment the pinned version in favor of the git URL once
# https://github.com/python-poetry/poetry-core/pull/257 is merged
# and a new poetry-core 1.0.x is released
# requires = ["poetry-core>=1.0.0"]
requires = ["poetry-core @ git+https://github.com/python-poetry/poetry-core.git@master"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-amazon-seller = 'tap_amazon_seller.tap:TapAmazonSeller.cli'
